Changed the popup system
The popup system is different now and also the warning system is active
