+++ Space-Zoologist/Assets/Prefabs/UI/GenericWindows/WarningWindow.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6e312741f118ff44d8914e9887ecdf5b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
