+++ Space-Zoologist/Assets/Scripts/DebuggingUI/SerializableNeed.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7dd54a3ce86aa454293f413947184206
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
