+++ Space-Zoologist/Assets/Scripts/DebuggingUI/SerializablePopulation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7a6466530a5b0b34eae99d11162868d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
