+++ Space-Zoologist/Assets/Scripts/DebuggingUI/VerboseInspector.cs
@@ -21,99 +21,6 @@ public class VerboseInspector : MonoBehaviour
             this.liquidBodies = new List<LiquidBody>(liquidBodies);
         }
     }
-    [System.Serializable]
-    public class NeedMetEntry
-    {
-        public string need;
-        public bool met;
-
-        public NeedMetEntry(KeyValuePair<NeedType, bool> entry)
-        {
-            need = entry.Key.ToString();
-            met = entry.Value;
-        }
-    }
-    [System.Serializable]
-    public class SerializableGrowthCalculator
-    {
-        public float maxFreshWaterTilePercent = GrowthCalculator.maxFreshWaterTilePercent;
-        public float maxSaltTilePercent = GrowthCalculator.maxSaltTilePercent;
-        public float maxBacteriaTilePercent = GrowthCalculator.maxBacteriaTilePercent;
-
-        public string growthStatus;
-        public int growthCountdown;
-        public int decayCountdown;
-        public float foodRating;
-        public float waterRating;
-        public float terrainRating;
-        public float populationIncreaseRate;
-        public List<NeedMetEntry> isNeedMet;
-
-        public SerializableGrowthCalculator(GrowthCalculator calculator)
-        {
-            growthStatus = calculator.GrowthStatus.ToString();
-            growthCountdown = calculator.GrowthCountdown;
-            decayCountdown = calculator.DecayCountdown;
-            foodRating = calculator.FoodRating;
-            waterRating = calculator.WaterRating;
-            terrainRating = calculator.TerrainRating;
-            populationIncreaseRate = calculator.populationIncreaseRate;
-
-            // Create a list with all need met entries
-            IEnumerable<NeedMetEntry> entries = calculator.IsNeedMet
-                .Select(entry => new NeedMetEntry(entry));
-            isNeedMet = new List<NeedMetEntry>(entries);
-        }
-    }
-    [System.Serializable]
-    public class SerializableNeed
-    {
-        public string needName;
-        public string needType;
-        public float needValue;
-        public float severity;
-        public bool isPreferred;
-
-        public SerializableNeed(Need need)
-        {
-            needName = need.NeedName;
-            needType = need.NeedType.ToString();
-            needValue = need.NeedValue;
-            severity = need.Severity;
-            isPreferred = need.IsPreferred;
-        }
-    }
-    [System.Serializable]
-    public class SerializablePopulation
-    {
-        public int count;
-        public int prePopulationCount;
-        public float foodDominance;
-        public Vector3 origin;
-        public List<SerializableNeed> needs;
-        public SerializableGrowthCalculator growthCalculator;
-        public bool isPaused;
-        public bool hasAccessibilityChanged;
-
-        public SerializablePopulation(Population population)
-        {
-            // Set the first few fields
-            count = population.Count;
-            prePopulationCount = population.PrePopulationCount;
-            foodDominance = population.FoodDominance;
-            origin = population.Origin;
-
-            // Create a list with all the entries in the needs dictionary
-            IEnumerable<SerializableNeed> entries = population.Needs.Values
-                .Select(need => new SerializableNeed(need));
-            needs = new List<SerializableNeed>(entries);
-
-            // Set the remaining fields
-            growthCalculator = new SerializableGrowthCalculator(population.GrowthCalculator);
-            isPaused = population.IsPaused;
-            hasAccessibilityChanged = population.HasAccessibilityChanged;
-        }
-    }
     #endregion
 
     #region Private Editor Fields
