+++ Space-Zoologist/Assets/Scripts/UI/WarningSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5304b04007bd4884da46d794f704fd9c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
