+++ Space-Zoologist/Assets/Scripts/Y - Utils/UI/GenericWindow/Deprecated/GenericWindowResponse.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public enum GenericWindowResponse
+{
+    Primary, Secondary
+}
