+++ Space-Zoologist/Assets/Scripts/Y - Utils/UI/GenericWindow/Deprecated/GenericWindowResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 85bfac267a110044dac8dc5c5955a83a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
