+++ Space-Zoologist/ProjectSettings/TagManager.asset
@@ -8,6 +8,7 @@ TagManager:
   - FoodSource
   - tiletodelete
   - Notebook
+  - MainCanvas
   layers:
   - Default
   - TransparentFX
